Update Plant Attributes' type

+++ src/main/java/plant/Cherry.java
@@ -6,7 +6,7 @@ public class Cherry extends Tree {
     }
 
     @Override
-    public void water(double amount) {
+    public void water(int amount) {
         System.out.println("Watering the cherry tree with " + amount + " units of water.");
     }
 }
