+++ src/main/java/plant/PlantType.java
@@ -5,19 +5,19 @@ import java.util.ArrayList;
 import java.util.List;
 
 public enum PlantType {
-    TREE(20.0, 100.0, 30.0, 32.0, 122.0, new ArrayList<>(Arrays.asList("Aphid", "Spider"))),
-    FLOWER(10.0, 50.0, 20.0, 32.0, 122.0, new ArrayList<>(Arrays.asList("Aphid", "Spider"))),
-    CROP(30.0, 80.0, 40.0, 32.0, 122.0, new ArrayList<>(Arrays.asList("Aphid", "Spider", "Whitefly")));
+    TREE(20, 100, 30, 32, 122, new ArrayList<>(Arrays.asList("Aphid", "Spider"))),
+    FLOWER(10, 50, 20, 32, 122, new ArrayList<>(Arrays.asList("Aphid", "Spider"))),
+    CROP(30, 80, 40, 32, 122, new ArrayList<>(Arrays.asList("Aphid", "Spider", "Whitefly")));
 
-    private final double minWaterLevel;
-    private final double maxWaterLevel;
-    private final double lowWaterThreshold;
-    private final double minTemperature;
-    private final double maxTemperature;
+    private final int minWaterLevel;
+    private final int maxWaterLevel;
+    private final int lowWaterThreshold;
+    private final int minTemperature;
+    private final int maxTemperature;
     private final List<String> pests;
 
-    PlantType(double minWaterLevel, double maxWaterLevel, double lowWaterThreshold,
-              double minTemperature, double maxTemperature, List<String> pests) {
+    PlantType(int minWaterLevel, int maxWaterLevel, int lowWaterThreshold,
+              int minTemperature, int maxTemperature, List<String> pests) {
         this.minWaterLevel = minWaterLevel;
         this.maxWaterLevel = maxWaterLevel;
         this.lowWaterThreshold = lowWaterThreshold;
@@ -26,23 +26,23 @@ public enum PlantType {
         this.pests = pests;
     }
 
-    public double getMinWaterLevel() {
+    public int getMinWaterLevel() {
         return minWaterLevel;
     }
 
-    public double getMaxWaterLevel() {
+    public int getMaxWaterLevel() {
         return maxWaterLevel;
     }
 
-    public double getLowWaterThreshold() {
+    public int getLowWaterThreshold() {
         return lowWaterThreshold;
     }
 
-    public double getMinTemperature() {
+    public int getMinTemperature() {
         return minTemperature;
     }
 
-    public double getMaxTemperature() {
+    public int getMaxTemperature() {
         return maxTemperature;
     }
