+++ src/main/java/plant/Rose.java
@@ -6,7 +6,7 @@ public class Rose extends Flower {
     }
 
     @Override
-    public void water(double amount) {
+    public void water(int amount) {
         System.out.println("Watering the rose with " + amount + " ml of water");
     }
 }
